refactor: remove file structure cruft
