+++ commands/ls.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 	"strings"
-	argparse "v/internal/argparse"
-	stateManager "v/internal/state"
+	argparse "v/argparse"
+	stateManager "v/state"
 )
 
 func ListVersions(args []string, flags argparse.Flags, currentState stateManager.State) error {
