+++ scripts/build
@@ -1,3 +1,3 @@
 #!/usr/bin/bash
 
-go build cmd/v.go
+go build .
